Move quests page into one component
